Add files via upload
This project is about an app that a user can read the Quran. I coded so that an AI will test your pronunciation when giving specific tasks; a chapter, verses or a verse. 
The reason for this project is our mothers language can have an affect of our pronunciation when reciting so this tests on this area.

FYI.... Still i works of improving the code...