Add files via upload
This file is the one I'm working on currently. So far, 7 languages are available to use. And their options of response types: 
- Audio - Repeats what you are saying depending on what language you want AI to speak.
- Type - You say something in your desired language, and it will repeat what you said depending on the language you select.
- Reply - You can have a conversation with AI based on what the desired language you select. 

Main purpose of this app is for people to learn and become better in the language they are pursuing. And of course, speaking is better than just writing!